made code cls compliant

+++ X10D.Performant.Tests/src/Core/BooleanTests.cs
@@ -3,13 +3,13 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace X10D.Performant.Tests.Core
 {
     /// <summary>
-    ///     Tests for <see cref="BooleanExtensions" />.
+    ///     Tests for <see cref="BooleanExtensions"/>.
     /// </summary>
     [TestClass]
     public class BooleanTests
     {
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToByte" />.
+        ///     Tests for <see cref="BooleanExtensions.ToByte"/>.
         /// </summary>
         [TestMethod]
         public void ToByte()
@@ -24,7 +24,7 @@ namespace X10D.Performant.Tests.Core
         }
 
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToDecimal" />.
+        ///     Tests for <see cref="BooleanExtensions.ToDecimal"/>.
         /// </summary>
         [TestMethod]
         public void ToDecimal()
@@ -39,22 +39,22 @@ namespace X10D.Performant.Tests.Core
         }      
         
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToDouble" />.
+        ///     Tests for <see cref="BooleanExtensions.ToDouble"/>.
         /// </summary>
         [TestMethod]
         public void ToDouble()
         {
             const bool a = true;
             const bool b = false;
-            const decimal c = 1;
-            const decimal d = 0;
+            const double c = 1;
+            const double d = 0;
 
             Assert.AreEqual(c, a.ToDouble());
             Assert.AreEqual(d, b.ToDouble());
         }
 
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToInt16" />.
+        ///     Tests for <see cref="BooleanExtensions.ToInt16"/>.
         /// </summary>
         [TestMethod]
         public void ToInt16()
@@ -69,7 +69,7 @@ namespace X10D.Performant.Tests.Core
         }
 
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToInt32" />.
+        ///     Tests for <see cref="BooleanExtensions.ToInt32"/>.
         /// </summary>
         [TestMethod]
         public void ToInt32()
@@ -84,7 +84,7 @@ namespace X10D.Performant.Tests.Core
         }
 
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToInt64" />.
+        ///     Tests for <see cref="BooleanExtensions.ToInt64"/>.
         /// </summary>
         [TestMethod]
         public void ToInt64()
@@ -99,7 +99,7 @@ namespace X10D.Performant.Tests.Core
         }
 
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToSByte" />.
+        ///     Tests for <see cref="BooleanExtensions.ToSByte"/>.
         /// </summary>
         [TestMethod]
         public void ToSByte()
@@ -114,7 +114,7 @@ namespace X10D.Performant.Tests.Core
         }      
         
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToSingle" />.
+        ///     Tests for <see cref="BooleanExtensions.ToSingle"/>.
         /// </summary>
         [TestMethod]
         public void ToFloat()
@@ -129,7 +129,7 @@ namespace X10D.Performant.Tests.Core
         }
 
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToUInt16" />.
+        ///     Tests for <see cref="BooleanExtensions.ToUInt16"/>.
         /// </summary>
         [TestMethod]
         public void ToUInt16()
@@ -144,7 +144,7 @@ namespace X10D.Performant.Tests.Core
         }
 
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToUInt32" />.
+        ///     Tests for <see cref="BooleanExtensions.ToUInt32"/>.
         /// </summary>
         [TestMethod]
         public void ToUInt32()
@@ -159,7 +159,7 @@ namespace X10D.Performant.Tests.Core
         }
 
         /// <summary>
-        ///     Tests for <see cref="BooleanExtensions.ToUInt64" />.
+        ///     Tests for <see cref="BooleanExtensions.ToUInt64"/>.
         /// </summary>
         [TestMethod]
         public void ToUInt64()
